Allow custom sqlite files
By exporting a function that gives access to the database, rather than a set of functions that are normally provided, developers can specify different database files to store data. Running the function allows you to set the file, which is json.sqlite if no file is provided, and will return the normal methods such as set, get, push, delete, etc. This would be great for large systems where it needs multiple files, or if there are multiple programs running in the same directory.